+++ backend/src/Controllers/ArrivalFlightsController.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using FlightApi.Models;
-using Interfaces.ArrivalFlightsService;
+using Interfaces.FlightService;
 
 
 
@@ -8,17 +8,17 @@ using Interfaces.ArrivalFlightsService;
 [ApiController]
 public class ArrivalFlightsController : ControllerBase
 {
-    private readonly IArrivalFlightsService arrivalFlightsService;
+    private readonly IFlightService flightService;
 
-    public ArrivalFlightsController(IArrivalFlightsService arrivalFlightsService){
-        this.arrivalFlightsService = arrivalFlightsService;
+    public ArrivalFlightsController(IFlightService flightService){
+        this.flightService = flightService;
     }
 
     // GET: ArrivalFlights/icao?icao=KSDF&toLoca=2022-10-05T08:00&fromLocal=2022-10-04T20:00
     [HttpGet("icao")]
     public async Task<ActionResult<Flights>> GetFlights(string toLocal, string fromLocal, string icao)
     {
-        var flights = await arrivalFlightsService.GetAllArrivalFlights(toLocal, fromLocal, icao);
+        var flights = await flightService.GetAllArrivalFlights(toLocal, fromLocal, icao);
         return flights;
     }
