+++ backend/src/Controllers/FlightsController.cs
@@ -1,6 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
 using FlightApi.Models;
-using Services.FlightService;
 using Interfaces.FlightService;
