--- backend/src/Interfaces/IArrivalFlightsService.cs
@@ -1,7 +1,0 @@
-using FlightApi.Models;
-
-namespace Interfaces.ArrivalFlightsService;
-public interface IArrivalFlightsService
-{
-    public Task<Flights> GetAllArrivalFlights(string toLocal, string fromLocal, string icao);
-}
